phase_3_rag_correction/requirements.txt
openai>=1.0.0
sentence-transformers>=2.2.0
torch>=1.9.0
numpy>=1.21.0
pyyaml>=6.0 